requirements.txt
nerodia
python-decouple